+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Feature
 - Bulk pseudonymization support adding unsure matches
 - Support importing policies from a csv file
+- Support language url parameter
 ## [1.0.3] - 2025-10-30
 ### Fix
 - Upgrade keycloak client to v25.0.6